+++ src/map.rs
@@ -1,6 +1,10 @@
 //! Non-empty [`HashMap`]s.
 
 use core::fmt;
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
+use crate::{FromNonEmptyIterator, IntoNonEmptyIterator, NonEmptyIterator};
 use std::borrow::Borrow;
 use std::collections::HashMap;
 use std::hash::BuildHasher;
@@ -40,7 +44,16 @@ macro_rules! nem {
 /// let m = nem!["elves" => 3000, "orcs" => 10000];
 /// assert_eq!(2, m.len().get());
 /// ```
-#[derive(Clone)]
+#[cfg_attr(
+    feature = "serde",
+    derive(Deserialize, Serialize),
+    serde(bound(
+        serialize = "K: Eq + Hash + Clone + Serialize, V: Clone + Serialize, S: Clone + BuildHasher",
+        deserialize = "K: Eq + Hash + Clone + Deserialize<'de>, V: Deserialize<'de>, S: Default + BuildHasher"
+    )),
+    serde(into = "HashMap<K, V, S>", try_from = "HashMap<K, V, S>")
+)]
+#[derive(Debug, Clone)]
 pub struct NEMap<K, V, S = std::collections::hash_map::RandomState> {
     inner: HashMap<K, V, S>,
 }
@@ -371,6 +384,40 @@ where
     }
 }
 
+impl<K, V, S> TryFrom<HashMap<K, V, S>> for NEMap<K, V, S>
+where
+    // FIXME: 2024-07-19 Fix the lower clone, thereby removing the need for `Clone` here.
+    K: Eq + Hash + Clone,
+    S: BuildHasher,
+{
+    type Error = crate::Error;
+
+    fn try_from(mut map: HashMap<K, V, S>) -> Result<Self, Self::Error> {
+        if map.is_empty() {
+            Err(crate::Error::Empty)
+        } else {
+            // NOTE 2024-07-19 These are safe unwraps due to the emptiness check
+            // above. It was done this way, instead of a `match`, in order to
+            // avoid an ownership problem.
+            let head_key = {
+                let k = map.keys().next().unwrap();
+                // FIXME: 2024-07-19 Avoid this clone.
+                k.clone()
+            };
+
+            let head_val = map.remove(&head_key).unwrap();
+
+            let ne = NEMap {
+                head_key,
+                head_val,
+                tail: map,
+            };
+
+            Ok(ne)
+        }
+    }
+}
+
 impl<K, V, S> IntoNonEmptyIterator for NEMap<K, V, S> {
     type IntoNEIter = IntoIter<K, V>;
 
@@ -579,12 +626,60 @@ impl<K: fmt::Debug, V: fmt::Debug, S> fmt::Debug for NEMap<K, V, S> {
 
 #[cfg(test)]
 mod test {
-    use maplit::hashmap;
+    use maplit::hashmap;    
+    use crate::nem;
+    use std::num::NonZeroUsize;
 
     #[test]
     fn debug_impl() {
         let expected = format!("{:?}", hashmap! {0 => 10});
         let actual = format!("{:?}", nem! {0 => 10});
         assert_eq!(expected, actual);
+
+
+    struct Foo {
+        user: String,
+    }
+
+    #[test]
+    fn macro_usage() {
+        let a = Foo {
+            user: "a".to_string(),
+        };
+        let b = Foo {
+            user: "b".to_string(),
+        };
+
+        let map = nem![1 => a, 2 => b];
+        assert_eq!("a", map.get(&1).unwrap().user);
+        assert_eq!("b", map.get(&2).unwrap().user);
+    }
+
+    #[test]
+    fn macro_length() {
+        let map = nem![1 => 'a', 2 => 'b', 1 => 'c'];
+        assert_eq!(unsafe { NonZeroUsize::new_unchecked(2) }, map.len());
+        assert_eq!('c', *map.get(&1).unwrap());
+        assert_eq!('b', *map.get(&2).unwrap());
+    }
+}
+
+#[cfg(feature = "serde")]
+#[cfg(test)]
+mod serde_tests {
+    use crate::{nem, NEMap};
+    use std::collections::HashMap;
+
+    #[test]
+    fn json() {
+        let map0 = nem![1 => 'a', 2 => 'b', 1 => 'c'];
+        let j = serde_json::to_string(&map0).unwrap();
+        let map1 = serde_json::from_str(&j).unwrap();
+        assert_eq!(map0, map1);
+
+        let empty: HashMap<usize, char> = HashMap::new();
+        let j = serde_json::to_string(&empty).unwrap();
+        let bad = serde_json::from_str::<NEMap<usize, char>>(&j);
+        assert!(bad.is_err());
     }
 }