+++ src/set.rs
@@ -1,6 +1,11 @@
 //! Non-empty Sets.
 
 use core::fmt;
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
+use crate::iter::NonEmptyIterator;
+use crate::{FromNonEmptyIterator, IntoIteratorExt, IntoNonEmptyIterator};
 use std::borrow::Borrow;
 use std::collections::HashSet;
 use std::hash::BuildHasher;
@@ -89,7 +94,16 @@ macro_rules! nes {
 ///
 /// As these methods are all "mutate-in-place" style and are difficult to
 /// reconcile with the non-emptiness guarantee.
-#[derive(Clone)]
+#[cfg_attr(
+    feature = "serde",
+    derive(Serialize, Deserialize),
+    serde(bound(
+        serialize = "T: Eq + Hash + Clone + Serialize, S: Clone + BuildHasher",
+        deserialize = "T: Eq + Hash + Deserialize<'de>, S: Default + BuildHasher"
+    )),
+    serde(into = "HashSet<T, S>", try_from = "HashSet<T, S>")
+)]
+#[derive(Debug, Clone)]
 pub struct NESet<T, S = std::collections::hash_map::RandomState> {
     inner: HashSet<T, S>,
 }
@@ -626,5 +640,42 @@ mod test {
         let expected = format!("{:?}", hashset! {0}.iter());
         let actual = format!("{:?}", nes! {0}.iter());
         assert_eq!(expected, actual);
+    }
+}
+
+impl<T, S> TryFrom<HashSet<T, S>> for NESet<T, S>
+where
+    T: Eq + Hash,
+    S: BuildHasher + Default,
+{
+    type Error = crate::Error;
+
+    fn try_from(set: HashSet<T, S>) -> Result<Self, Self::Error> {
+        let ne = set
+            .try_into_nonempty_iter()
+            .ok_or(crate::Error::Empty)?
+            .collect();
+
+        Ok(ne)
+    }
+}
+
+#[cfg(feature = "serde")]
+#[cfg(test)]
+mod serde_tests {
+    use crate::{nes, NESet};
+    use std::collections::HashSet;
+
+    #[test]
+    fn json() {
+        let set0 = nes![1, 1, 2, 3, 2, 1, 4];
+        let j = serde_json::to_string(&set0).unwrap();
+        let set1 = serde_json::from_str(&j).unwrap();
+        assert_eq!(set0, set1);
+
+        let empty: HashSet<usize> = HashSet::new();
+        let j = serde_json::to_string(&empty).unwrap();
+        let bad = serde_json::from_str::<NESet<usize>>(&j);
+        assert!(bad.is_err());
     }
 }